Added full Admin Alert Functionality.
To do an admin alert, enter the realtime database on our project. Click the plus sign and then add message in the first box, and then in the value box put your message.
Added sourcing for android studio page for alerts.
Added sourcing in documents for wav file used for alerts.