Created username storage and retrieval. Bug possibility.
if you are trying to retrieve a username, do usernameStorage.username to access the variable.

There might be a bug with the code and the ui not responding. I've had a couple of times and error pop up but the most recent test on my Pixle 4 30 api phone had no issues.