Updated Alarm And Redid Positive Notifications.
Made it such that alarm has a new sound, and positive notifications now uses a thread system to play notifications in the background. I also made it such that it pulls from the database the affirmations and knows how many entries are in the database so we shouldn't need to change any code within the positive notification manager if we are to add more affirmations.