Updated Alarm Clock
Got rid of hidden buttons on page. Updated alarm such that it now takes an am or pm variable. No more military time. Also made it so that the buttons are no longer editable when you enter in an alarm. Edited toast such that it shows am or pm and proper times.